Verify scheduled job execution after changes to job definitions and schedules
