+++ functional-test/src/test/groovy/org/rundeck/util/common/projects/ProjectUtils.groovy
@@ -25,7 +25,7 @@ class ProjectUtils {
                 }
             }
             (1..jobsCountsPerProject).each {it2 ->
-                def pathFile = JobUtils.updateJobFileToImport("api-test-executions-running-scheduled.xml", projectName)
+                def pathFile = JobUtils.updateJobFileToImport("api-test-executions-running-scheduled.xml", projectName, ["args":"echo 1; sleep 5"])
                 def imported = JobUtils.jobImportFile(projectName, pathFile, client)
                 Thread.sleep(2000)
             }
