+++ functional-test/src/test/resources/test-files/api-test-executions-running-scheduled.xml
@@ -1,11 +1,11 @@
 <joblist>
   <job>
     <defaultTab>nodes</defaultTab>
-    <description></description>
+    <description>xml-job-description-name</description>
     <executionEnabled>true</executionEnabled>
     <id>d9500676-1cba-4e67-adec-1c28ad935048</id>
     <loglevel>INFO</loglevel>
-    <name>Simple Scheduled job</name>
+    <name>xml-job-name</name>
     <nodeFilterEditable>false</nodeFilterEditable>
     <plugins />
     <schedule crontab="xml-schedule-crontab"/>
@@ -13,7 +13,7 @@
     <multipleExecutions>true</multipleExecutions>
     <sequence keepgoing='false'>
       <command>
-        <exec>echo '1'; sleep 5</exec>
+        <exec>xml-args</exec>
       </command>
     </sequence>
     <uuid>xml-uuid</uuid>
